requirements.txt
asyncio
azure-ai-inference==1.0.0b8
azure-ai-projects==1.0.0b5
azure-ai-ml
azure-core==1.32.0
azure-identity
debugpy==1.8.12
dotenv
jsonschema==4.23.0
jsonschema-path==0.3.3
jsonschema-specifications==2024.10.1
jwt
ldap3
logging
msal==1.31.1
openai==1.60.0
pydantic==2.10.5
python-dotenv
rich
requests==2.32.3
semantic-kernel==1.19.0
streamlit
threading
time
traceback
typing